test: fold collision help coverage into selection tests
